+++ travis-run.sh
@@ -36,6 +36,16 @@ start_es6() {
   # curl http://127.0.0.1:$es_port && echo "ES is up!" || cat /tmp/$es_cluster.log ./elasticsearch/logs/$es_cluster.log
 }
 
+start_es7() {
+  es_args=$1
+  path_env=$2
+  es_port=$3
+  es_cluster=$4
+  ES_PATH_CONF=$path_env elasticsearch/bin/elasticsearch $es_args > /tmp/$es_cluster.log &
+  sleep 20
+  curl http://127.0.0.1:$es_port && echo "$es_cluster Elasticsearch is up!" || cat /tmp/$es_cluster.log ./elasticsearch/logs/$es_cluster.log
+}
+
 common_node_settings() {
   major=$1
   minor=$2
@@ -45,8 +55,16 @@ common_node_settings() {
   echo 'network.host: 127.0.0.1' > $file
   echo "http.port: ${port}" >> $file
   echo "cluster.name: ${clustername}" >> $file
+  echo "node.name: ${clustername}" >> $file
   echo 'node.max_local_storage_nodes: 2' >> $file
-  echo "discovery.zen.ping.unicast.hosts: [\"127.0.0.1:${port}\"]" >> $file
+  if [[ $major -lt 7 ]]; then
+    echo "discovery.zen.ping.unicast.hosts: [\"127.0.0.1:${port}\"]" >> $file
+  else
+    transport=$(($port+100))
+    echo "transport.port: ${transport}" >> $file
+    echo "discovery.seed_hosts: [\"localhost:${transport}\"]" >> $file
+    echo "discovery.type: single-node" >> $file
+  fi
   if [[ $major -ge 6 ]] && [[ $minor -ge 3 ]]; then
     echo 'xpack.monitoring.enabled: false' >> $file
     echo 'node.ml: false' >> $file
@@ -83,9 +101,12 @@ common_node_settings $MAJORVER $MINORVER 9201 remote "$RC/elasticsearch.yml"
 if [[ $MAJORVER -lt 6 ]]; then
   start_es $java_home "-d -Epath.conf=$LC" 9200 "local"
   start_es $java_home "-d -Epath.conf=$RC" 9201 "remote"
-else
-  start_es6 $java_home "-d" "$LC" 9200 "local"
-  start_es6 $java_home "-d" "$RC" 9201 "remote"
+elif [[ $MARJORVER -eq 6 ]]; then
+  start_es6 $java_home " " "$LC" 9200 "local"
+  start_es6 $java_home " " "$RC" 9201 "remote"
+else 
+  start_es7 " " "$LC" 9200 "local"
+  start_es7 " " "$RC" 9201 "remote"
 fi
 
 python setup.py test